@@ -6,31 +6,31 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.15.1 <3.0.0"
+  sdk: ">=2.18.2 <3.0.0"
 
 dependencies:
+  cupertino_icons: ^1.0.5
+  dio: ^4.0.6
   flutter:
     sdk: flutter
-  cupertino_icons: ^1.0.2
   flutter_iconly: ^1.0.2
-  ionicons: ^0.1.2
-  google_fonts: 2.2.0
-  dio: ^4.0.4
-  url_launcher: ^6.0.20
-  pull_to_refresh: ^2.0.0
-  intl: ^0.17.0
-  share_plus: ^3.1.0
-  hive: ^2.0.6
-  like_button: ^2.0.4
+  google_fonts: ^3.0.1
+  hive: ^2.2.3
   hive_flutter: ^1.1.0
+  intl: ^0.17.0
+  ionicons: ^0.2.1
+  like_button: ^2.0.5
+  pull_to_refresh: ^2.0.0
+  share_plus: ^6.1.0
+  url_launcher: ^6.1.6
 
 dev_dependencies:
+  build_runner: ^2.3.2
+  flutter_launcher_icons: ^0.10.0
+  flutter_lints: ^2.0.1
   flutter_test:
     sdk: flutter
-  flutter_lints: ^1.0.0
-  build_runner: ^2.1.8
-  hive_generator: ^1.1.2
-  flutter_launcher_icons: ^0.9.2
+  hive_generator: ^2.0.0
   rename: ^2.0.1
 
 flutter_icons:
@@ -48,7 +48,7 @@ flutter:
   assets:
     - assets/icon/icon.png
     - assets/icon/foreground.png
-
+    
   # fonts:
   #   - family: Schyler
   #     fonts:
