@@ -1,7 +1,7 @@
 import 'dart:convert';
+import 'dart:developer';
 
 import 'package:dio/dio.dart';
-import 'package:intl/intl.dart';
 import 'package:test_technique/environment.dart';
 import 'package:test_technique/models/news.dart';
 import 'package:test_technique/models/source.dart';
@@ -9,25 +9,38 @@ import 'package:test_technique/models/source.dart';
 class NewsService {
 
   final String apiKey = Environment.newsApiKeys;
+  final String baseUrl = "https://newsapi.org/v2";
+
+  Future<Response> get(String pathUrl) async {
+
+    Dio dio = Dio();
+    Response response = Response(requestOptions: RequestOptions(path: ""));
+
+    try {
+      response = await dio.get("$baseUrl$pathUrl&apiKey=$apiKey"); 
+      log("API Response: ${response.statusCode} ${response.statusMessage}");
+    } on DioError catch (e) {
+      if(e.response != null){
+        log("API Response: ${e.response!.statusCode} ${e.response!.statusMessage}");
+        Map errorData = e.response!.data;
+        if(errorData.containsKey("message")){
+          log("Error Response message: ${errorData["message"]}");
+        }
+      } else {
+        log("Error Response message: ${e.message}");
+      }
+    }
+
+    return response;
+  }
 
   Future<List<News>> getAllFrenchNews() async {
     final List<News> allNews = [];
-    Dio dio = Dio();
-    Response response = await dio.get("https://newsapi.org/v2/top-headlines?country=fr&apiKey=$apiKey");
+    Response response = await get("/top-headlines?country=fr");
     String body = response.toString();
-    // log(body);
     Map data = jsonDecode(body);
     for (int i = 0; i < data["articles"].length; i++) {
-      final datePublished = DateFormat('dd LLLL yyyy HH:mm',"FR").format(DateTime.parse(data["articles"][i]["publishedAt"]));
-      final News newsData = News(
-        data["articles"][i]["title"],
-        data["articles"][i]["author"],
-        data["articles"][i]["source"]["name"],
-        data["articles"][i]["url"],
-        data["articles"][i]["urlToImage"],
-        datePublished,
-        data["articles"][i]["content"]
-      );
+      final News newsData = News.fromJson(data["articles"][i]);
       allNews.add(newsData);
     }
 
@@ -36,22 +49,11 @@ class NewsService {
 
   Future<List<News>> getAllFrenchNewsBySource(Source source) async {
     final List<News> allNews = [];
-    Dio dio = Dio();
-    Response response = await dio.get("https://newsapi.org/v2/top-headlines?sources=${source.id}&apiKey=$apiKey");
+    Response response = await get("/top-headlines?sources=${source.id}");
     String body = response.toString();
-    // log(body);
     Map data = jsonDecode(body);
     for (int i = 0; i < data["articles"].length; i++) {
-      final datePublished = DateFormat('dd LLLL yyyy HH:mm',"FR").format(DateTime.parse(data["articles"][i]["publishedAt"]));
-      final News newsData = News(
-        data["articles"][i]["title"],
-        data["articles"][i]["author"],
-        data["articles"][i]["source"]["name"],
-        data["articles"][i]["url"],
-        data["articles"][i]["urlToImage"],
-        datePublished,
-        data["articles"][i]["content"]
-      );
+      final News newsData = News.fromJson(data["articles"][i]);
       allNews.add(newsData);
     }
 
@@ -60,22 +62,11 @@ class NewsService {
 
   Future<List<News>> getAllNewsByTitleSearch(String query) async {
     final List<News> allNews = [];
-    Dio dio = Dio();
-    Response response = await dio.get("https://newsapi.org/v2/everything?q=$query&apiKey=$apiKey");
+    Response response = await get("/everything?q=$query");
     String body = response.toString();
-    // log(body);
     Map data = jsonDecode(body);
     for (int i = 0; i < data["articles"].length; i++) {
-      final datePublished = DateFormat('dd LLLL yyyy HH:mm',"FR").format(DateTime.parse(data["articles"][i]["publishedAt"]));
-      final News newsData = News(
-        data["articles"][i]["title"],
-        data["articles"][i]["author"],
-        data["articles"][i]["source"]["name"],
-        data["articles"][i]["url"],
-        data["articles"][i]["urlToImage"],
-        datePublished,
-        data["articles"][i]["content"]
-      );
+      final News newsData = News.fromJson(data["articles"][i]);
       allNews.add(newsData);
     }
 
@@ -84,22 +75,11 @@ class NewsService {
 
   Future<List<News>> getAllFrenchNewsByCategorySearch(String query) async {
     final List<News> allNews = [];
-    Dio dio = Dio();
-    Response response = await dio.get("https://newsapi.org/v2/top-headlines?country=fr&category=$query&apiKey=$apiKey");
+    Response response = await get("/top-headlines?country=fr&category=$query");
     String body = response.toString();
-    // log(body);
     Map data = jsonDecode(body);
     for (int i = 0; i < data["articles"].length; i++) {
-      final datePublished = DateFormat('dd LLLL yyyy HH:mm',"FR").format(DateTime.parse(data["articles"][i]["publishedAt"]));
-      final News newsData = News(
-        data["articles"][i]["title"],
-        data["articles"][i]["author"],
-        data["articles"][i]["source"]["name"],
-        data["articles"][i]["url"],
-        data["articles"][i]["urlToImage"],
-        datePublished,
-        data["articles"][i]["content"]
-      );
+      final News newsData = News.fromJson(data["articles"][i]);
       allNews.add(newsData);
     }
 
@@ -108,24 +88,14 @@ class NewsService {
 
   Future<List<News>> getAllNewsBySourceSearch(String query) async {
     final List<News> allNews = [];
-    Dio dio = Dio();
+    
     final Source? source = await getSourceBySearch(query);
     if(source != null){
-      Response response = await dio.get("https://newsapi.org/v2/top-headlines?sources=${source.id}&apiKey=$apiKey");
+      Response response = await get("/top-headlines?sources=${source.id}");
       String body = response.toString();
-      // log(body);
       Map data = jsonDecode(body);
       for (int i = 0; i < data["articles"].length; i++) {
-        final datePublished = DateFormat('dd LLLL yyyy HH:mm',"FR").format(DateTime.parse(data["articles"][i]["publishedAt"]));
-        final News newsData = News(
-          data["articles"][i]["title"],
-          data["articles"][i]["author"],
-          data["articles"][i]["source"]["name"],
-          data["articles"][i]["url"],
-          data["articles"][i]["urlToImage"],
-          datePublished,
-          data["articles"][i]["content"]
-        );
+        final News newsData = News.fromJson(data["articles"][i]);
         allNews.add(newsData);
       }
     }
@@ -134,20 +104,12 @@ class NewsService {
 
   Future<List<Source>> getAllFrenchSources() async {
     final List<Source> allSources = [];
-    Dio dio = Dio();
-    Response response = await dio.get("https://newsapi.org/v2/top-headlines/sources?country=fr&apiKey=$apiKey");
+    
+    Response response = await get("/top-headlines/sources?country=fr");
     String body = response.toString();
-    // log(body);
     Map data = jsonDecode(body);
     for (int i = 0; i < data["sources"].length; i++) {
-      final Source sourceData = Source(
-        data["sources"][i]["name"],
-        data["sources"][i]["id"],
-        data["sources"][i]["description"],
-        data["sources"][i]["url"],
-        data["sources"][i]["category"],
-        data["sources"][i]["country"],
-      );
+      final Source sourceData = Source.fromJson(data["sources"][i]);
       allSources.add(sourceData);
     }
     return allSources;
@@ -156,20 +118,12 @@ class NewsService {
   Future<Source?> getSourceBySearch(String query) async {
     final List<Source> allSources = [];
     Source? finalSource;
-    Dio dio = Dio();
-    Response response = await dio.get("https://newsapi.org/v2/top-headlines/sources?apiKey=$apiKey");
+    
+    Response response = await get("/top-headlines/sources?");
     String body = response.toString();
-    // log(body);
     Map data = jsonDecode(body);
     for (int i = 0; i < data["sources"].length; i++) {
-      final Source sourceData = Source(
-        data["sources"][i]["name"],
-        data["sources"][i]["id"],
-        data["sources"][i]["description"],
-        data["sources"][i]["url"],
-        data["sources"][i]["category"],
-        data["sources"][i]["country"],
-      );
+      final Source sourceData = Source.fromJson(data["sources"][i]);
       allSources.add(sourceData);
     }
     for (final Source source in allSources) {
