@@ -6,7 +6,7 @@ import FlutterMacOS
 import Foundation
 
 import path_provider_macos
-import share_plus_macos
+import share_plus
 import url_launcher_macos
 
 func RegisterGeneratedPlugins(registry: FlutterPluginRegistry) {
